# [Go] Use flatten slice for the custom pQueue