--- a/go/main.go
+++ b/go/main.go
@@ -68,31 +68,34 @@ func main() {
 			}
 		}
 		taggedPostCount[i] = 0 // Don't count self
-		top5 := [topN]PostWithSharedTags{}
+		top5 := [topN * 2]int{}
 		minTags := byte(0)
 
 		for j, count := range taggedPostCount {
 			if count > minTags {
-				pos := 3
-				for pos >= 0 && top5[pos].SharedTags < count {
-					pos--
-				}
-				pos++
+				upperBound := (topN - 2) * 2
 
-				// Shift and insert
-				if pos < 4 {
-					copy(top5[pos+1:], top5[pos:4])
+				countInt := int(count)
+				for upperBound >= 0 && countInt > top5[upperBound] {
+					top5[upperBound+2] = top5[upperBound]
+					top5[upperBound+3] = top5[upperBound+1]
+					upperBound -= 2
 				}
 
-				top5[pos] = PostWithSharedTags{Post: isize(j), SharedTags: count}
-				minTags = top5[4].SharedTags
+				insertPos := upperBound + 2
+				top5[insertPos] = countInt
+				top5[insertPos+1] = j
+
+				minTags = byte(top5[topN*2-2])
 			}
 		}
 
 		// Convert indexes back to Post pointers
 		topPosts := [topN]*Post{}
-		for idx, t := range top5 {
-			topPosts[idx] = &posts[t.Post]
+
+		for j := 0; j < topN*2; j += 2 {
+			index := top5[j+1]
+			topPosts[j/2] = &posts[index]
 		}
 
 		allRelatedPosts[i] = RelatedPosts{