Main:  `22.65 ms  |  322.43 ms  |  2.86 s`

This PR:
```
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 21.69926ms
        Processing time (w/o IO): 21.683359ms
        Processing time (w/o IO): 21.522457ms
        Processing time (w/o IO): 21.613658ms
        Processing time (w/o IO): 22.231371ms
        Processing time (w/o IO): 21.630259ms
        Processing time (w/o IO): 21.587957ms
        Processing time (w/o IO): 22.033167ms
        Processing time (w/o IO): 21.656059ms
        Processing time (w/o IO): 23.635201ms
        Processing time (w/o IO): 22.215771ms
        Processing time (w/o IO): 21.878764ms
        Processing time (w/o IO): 21.605458ms
          Time (mean ± σ):      61.0 ms ±   1.4 ms    [User: 59.4 ms, System: 8.6 ms]
          Range (min … max):    58.0 ms …  62.7 ms    10 runs
         
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 307.217011ms
        Processing time (w/o IO): 310.122574ms
        Processing time (w/o IO): 310.89489ms
          Time (mean ± σ):     451.9 ms ±   3.0 ms    [User: 432.4 ms, System: 41.9 ms]
          Range (min … max):   449.8 ms … 453.9 ms    2 runs
         
Go:

        Benchmark 1: ./related
        Processing time (w/o IO): 2.718751456s
        Processing time (w/o IO): 2.707466583s
        Processing time (w/o IO): 2.721011872s
          Time (mean ± σ):      3.194 s ±  0.019 s    [User: 3.215 s, System: 0.104 s]
          Range (min … max):    3.180 s …  3.207 s    2 runs
```